+++ src/win32/i_system.cpp
@@ -1423,11 +1423,11 @@ static HCURSOR CreateBitmapCursor(int xhot, int yhot, HBITMAP and_mask, HBITMAP 
 {
 	ICONINFO iconinfo =
 	{
-		FALSE,		// fIcon
-		xhot,		// xHotspot
-		yhot,		// yHotspot
-		and_mask,	// hbmMask
-		color_mask	// hbmColor
+		FALSE,			// fIcon
+		(DWORD)xhot,	// xHotspot
+		(DWORD)yhot,	// yHotspot
+		and_mask,		// hbmMask
+		color_mask		// hbmColor
 	};
 	HCURSOR cursor = CreateIconIndirect(&iconinfo);
